Bootstrap 4 beta compatibility. Update priv policy. Header btns changes.
Bootstrap 4, popper.js, and jquery use cdn to load. Jquery is still a package dependency for Toastr in npm, so it stays in package.json. "About" page collapsible items now work again. It was broken because jquery needed to be linked in index.html. Home and Sign Out buttons now restrained by container.